Fix scheduler

+++ config/initializers/scheduler.rb
@@ -1,23 +1,30 @@
 require 'rufus-scheduler'
 
-scheduler = Rufus::Scheduler.new
+if (
+  ! defined?(Rails::Console) &&
+  File.basename($0) != 'rake' &&
+  Rails.configuration.respond_to?(:scheduler)
+) then
 
-scheduler.in '1h' do
-  # Delete all prev posts
-  FileUtils.rm_rf("#{Rails.root.join('web_posts')}/.", secure: true)
-  # Generate new posts
-  FB.generate_web_post_from_page_post
-  # Move new posts to web
-  FileUtils.cp_r("#{Rails.root.join('web_posts')}/.", "#{ENV['PLAYQPID_POST_FOLDER_PATH']}/")
-  cmd = <<~EOF
-    cd #{ENV['PLAYQPID_FOLDER_PATH']}
-    git add .
-    git commit -m "Add posts"
-    git push heroku master
-  EOF
-  `#{cmd}`
+  scheduler = Rufus::Scheduler.new
 
-  FileUtils.rm_rf("#{Rails.root.join('web_posts')}/.", secure: true)
+  scheduler.in '1h' do
+    # Delete all prev posts
+    FileUtils.rm_rf("#{Rails.root.join('web_posts')}/.", secure: true)
+    # Generate new posts
+    FB.generate_web_post_from_page_post
+    # Move new posts to web
+    FileUtils.cp_r("#{Rails.root.join('web_posts')}/.", "#{ENV['PLAYQPID_POST_FOLDER_PATH']}/")
+    cmd = <<~EOF
+      cd #{ENV['PLAYQPID_FOLDER_PATH']}
+      git add .
+      git commit -m "Add posts"
+      git push heroku master
+    EOF
+    `#{cmd}`
+
+    FileUtils.rm_rf("#{Rails.root.join('web_posts')}/.", secure: true)
+  end
+
+  scheduler.join
 end
-
-scheduler.join
